+++ hexital/exceptions.py
@@ -11,3 +11,8 @@ class InvalidPattern(Exception):
 class InvalidTimeFrame(Exception):
     def __init__(self, message):
         super().__init__(message)
+
+
+class InvalidCandleOrder(Exception):
+    def __init__(self, message):
+        super().__init__(message)