+++ hexital/lib/timeframe_utils.py
@@ -6,10 +6,26 @@ VALID_TIMEFRAME_PREFIXES = ["S", "T", "H", "D"]
 
 
 def round_down_timestamp(timestamp: datetime, timeframe: timedelta) -> datetime:
-    timestamp = timestamp.replace(microsecond=0)
+    """Find and round down timestamp to the nearest matching timeframe. E.G timeframe of 5 minute
+    E.G T5: 09:00:01 -> 9:00:00
+    E.G T5: 09:01:20 -> 9:00:00
+    E.G T5: 09:05:00 -> 9:05:00
+    Note: This method also calls clean_timestamp, removing microseconds
+    """
+    timestamp = clean_timestamp(timestamp)
     return datetime.fromtimestamp(
         timestamp.timestamp() // timeframe.total_seconds() * timeframe.total_seconds()
     )
+
+
+def on_timeframe(timestamp: datetime, timeframe: timedelta) -> bool:
+    """Checks if timestamp is on a timeframe value"""
+    return timestamp.timestamp() % timeframe.total_seconds() == 0
+
+
+def clean_timestamp(timestamp: datetime) -> datetime:
+    """Removes Microseconds from the timestamp and returns it"""
+    return timestamp.replace(microsecond=0)
 
 
 def timeframe_to_timedelta(timeframe: str) -> timedelta: