+++ hexital/lib/__init__.py
@@ -1,1 +1,2 @@
-from . import candle_extension, timeframe_utils, utils # noqa F401
+from . import candle_extension, timeframe_utils, utils  # noqa F401
+from .timeframe_utils import TimeFrame