+++ tests/lib/test_timeframe_utils.py
@@ -3,6 +3,7 @@ from datetime import datetime, timedelta
 import pytest
 from hexital.exceptions import InvalidTimeFrame
 from hexital.lib.timeframe_utils import round_down_timestamp, timeframe_to_timedelta
+from hexital import TimeFrame
 
 
 def test_timeframe_to_delta_min():
@@ -35,10 +36,18 @@ def test_timeframe_to_delta_no_key():
         assert timeframe_to_timedelta("30")
 
 
+def test_timeframe_enum_to_delta_min():
+    assert timeframe_to_timedelta(TimeFrame.MINUTE10) == timedelta(minutes=10)
+
+
+def test_timeframe_enum_to_delta_secs():
+    assert timeframe_to_timedelta(TimeFrame.SECOND10) == timedelta(seconds=10)
+
+
 def test_round_down_minutes_no_change():
-    assert round_down_timestamp(
-        datetime(2023, 6, 6, 12, 0, 0), timedelta(minutes=1)
-    ) == datetime(2023, 6, 6, 12, 0, 0)
+    assert round_down_timestamp(datetime(2023, 6, 6, 12, 0, 0), timedelta(minutes=1)) == datetime(
+        2023, 6, 6, 12, 0, 0
+    )
 
 
 def test_round_down_minutes_remove_seconds_and_mili():