+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 - Cleaned up code and some potential Bugs ruff/pyright
 - Added movement Above/Below and updated others to use it
 - Updated collapse candle 'fill' to show essentially doji candle rather than copy prev
+- Major Fix: Re-wrote collapse_candles_timeframe to correctly handle candles,gaps and appending
 - Fixed Timeframe bug with candles reference in indicator that use sub indicators
 - Fixed possible error in VWAP with no volumes traded
 - Fixed bug with nested Indicator returning None for valid 0 Value