+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 - Added movement Above/Below and updated others to use it
 - Updated collapse candle 'fill' to show essentially doji candle rather than copy prev
 - Major Fix: Re-wrote collapse_candles_timeframe to correctly handle candles,gaps and appending
+- Fixed Doji pattern
 - Fixed Timeframe bug with candles reference in indicator that use sub indicators
 - Fixed possible error in VWAP with no volumes traded
 - Fixed bug with nested Indicator returning None for valid 0 Value